server/requirements.txt
flask==2.2.2
sseclient-py==1.7.2
colorama==0.4.5
